+++ src/DynamicPPL.jl
@@ -186,7 +186,7 @@ include("context_implementations.jl")
 include("compiler.jl")
 include("pointwise_logdensities.jl")
 include("submodel_macro.jl")
-include("test_utils.jl")
+include("test_utils/main.jl")
 include("transforming.jl")
 include("logdensityfunction.jl")
 include("model_utils.jl")
